feat(autocmds): add basic autocmds wrapper
